backend/requirements/development.txt
# DEVELOPMENT ENVIRONMENT REQUIREMENTS
# =====================================

# Include the base requirements
-r base.txt

# Code style
autopep8==1.5.4

# Debug
django-debug-toolbar==3.2.1

# Shell
ipython==7.16.1
db-sqlite3==0.0.1
